+++ spec/spec_helper.rb
@@ -10,6 +10,8 @@ require "whatweb"
 require "vcr"
 require "webmock"
 
+require_relative "./support/helpers/helpers"
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
@@ -21,17 +23,7 @@ RSpec.configure do |config|
     c.syntax = :expect
   end
 
-  def capture(stream)
-    begin
-      stream = stream.to_s
-      eval "$#{stream} = StringIO.new"
-      yield
-      result = eval("$#{stream}").string
-    ensure
-      eval("$#{stream} = #{stream.upcase}")
-    end
-    result
-  end
+  config.include Spec::Support::Helpers
 end
 
 VCR.configure do |config|
