+++ spec/support/helpers/helpers.rb
@@ -1,0 +1,17 @@
+module Spec
+  module Support
+    module Helpers
+      def capture(stream)
+        begin
+          stream = stream.to_s
+          eval "$#{stream} = StringIO.new"
+          yield
+          result = eval("$#{stream}").string
+        ensure
+          eval("$#{stream} = #{stream.upcase}")
+        end
+        result
+      end
+    end
+  end
+end
