+++ spec/whatweb_spec.rb
@@ -6,17 +6,15 @@ RSpec.describe WhatWeb, vcr: vcr_options do
   it "has a version number" do
     expect(subject::VERSION).not_to be nil
   end
-  context "without the aggressive mode" do
-    describe "#execute_plugins" do
+  describe ".execute_plugins" do
+    context "without the aggressive mode" do
       it "should return a Hash" do
         results = subject.execute_plugins "https://www.webscantest.com"
         expect(results).to be_a(Hash)
         expect(results.dig("Title").first.dig(:string)).to eq("Test Site")
       end
     end
-  end
-  context "with the aggressive mode" do
-    describe "#execute_plugins" do
+    context "with the aggressive mode" do
       it "should return a Hash" do
         allow_any_instance_of(WhatWeb::Plugin).to receive(:randstr).and_return("fqeewoohxevinxslnplhjbymmeplmkwl")
         results = subject.execute_plugins("https://www.webscantest.com", is_aggressive: true)
@@ -24,7 +22,7 @@ RSpec.describe WhatWeb, vcr: vcr_options do
       end
     end
   end
-  describe "#plugin_names" do
+  describe ".plugin_names" do
     it "should return an Array " do
       array = subject.plugin_names
       expect(array).to be_an(Array)
